Extract TypeScript project generation into function, simplify path imports
