+++ src/bundler/angular-package.bundler.ts
@@ -1,4 +1,4 @@
-import { posix as path } from 'path';
+import * as path from 'path';
 
 import * as semver from 'semver';
 import { rollup, OutputOptions, RollupFileOptions, Bundle } from 'rollup';
