+++ src/bundler/rollup-configuration-builder.ts
@@ -1,4 +1,4 @@
-import { posix as path } from 'path';
+import * as path from 'path';
 
 import { OutputOptions, RollupWarning, RollupFileOptions } from 'rollup';
 import * as rollupCommonjsPlugin from 'rollup-plugin-commonjs';
