+++ src/transformer/angular-package.transformer.ts
@@ -1,4 +1,4 @@
-import { posix as path } from 'path';
+import * as path from 'path';
 
 import Project, { SourceFile } from 'ts-simple-ast';
