+++ src/transformer/external-resources/angular-external-resources.file-analyzer.ts
@@ -1,4 +1,4 @@
-import { posix as path } from 'path';
+import * as path from 'path';
 
 import { ObjectLiteralExpression, ClassDeclaration, SourceFile, Decorator, PropertyAssignment, StringLiteral } from 'ts-simple-ast';
